Moved DesignedForm to an own forms module.
