+++ form_designer/forms.py
@@ -1,0 +1,17 @@
+from django import forms
+from django.forms import widgets
+
+class DesignedForm(forms.Form):
+    def __init__(self, form_definition, initial_data=None, *args, **kwargs):
+        super(DesignedForm, self).__init__(*args, **kwargs)
+        for def_field in form_definition.formdefinitionfield_set.all():
+            self.add_defined_field(def_field, initial_data)
+        self.fields[form_definition.submit_flag_name] = forms.BooleanField(required=False, initial=1, widget=widgets.HiddenInput)
+
+    def add_defined_field(self, def_field, initial_data=None):
+        if initial_data and initial_data.has_key(def_field.name):
+            if not def_field.field_class in ('django.forms.MultipleChoiceField', 'django.forms.ModelMultipleChoiceField'):
+                def_field.initial = initial_data.get(def_field.name)
+            else:
+                def_field.initial = initial_data.getlist(def_field.name)
+        self.fields[def_field.name] = get_class(def_field.field_class)(**def_field.get_form_field_init_args())
