+++ form_designer/views.py
@@ -1,5 +1,4 @@
 from django import forms
-from django.forms import widgets
 from django.shortcuts import get_object_or_404, render_to_response
 from django.template import RequestContext
 from django.utils.translation import ugettext as _
@@ -8,22 +7,8 @@ from django.conf import settings
 from django.contrib import messages
 from django.core.context_processors import csrf
 
+from form_designer.forms import DesignedForm
 from form_designer.models import FormDefinition, get_class
-
-class DesignedForm(forms.Form):
-    def __init__(self, form_definition, initial_data=None, *args, **kwargs):
-        super(DesignedForm, self).__init__(*args, **kwargs)
-        for def_field in form_definition.formdefinitionfield_set.all():
-            self.add_defined_field(def_field, initial_data)
-        self.fields[form_definition.submit_flag_name] = forms.BooleanField(required=False, initial=1, widget=widgets.HiddenInput)
-
-    def add_defined_field(self, def_field, initial_data=None):
-        if initial_data and initial_data.has_key(def_field.name):
-            if not def_field.field_class in ('django.forms.MultipleChoiceField', 'django.forms.ModelMultipleChoiceField'):
-                def_field.initial = initial_data.get(def_field.name)
-            else:
-                def_field.initial = initial_data.getlist(def_field.name)
-        self.fields[def_field.name] = get_class(def_field.field_class)(**def_field.get_form_field_init_args())
 
 def process_form(request, form_definition, context={}, is_cms_plugin=False):
     success_message = form_definition.success_message or _('Thank you, the data was submitted successfully.')
@@ -80,8 +65,8 @@ def detail(request, object_name):
     result = process_form(request, form_definition)
     if isinstance(result, HttpResponseRedirect):
         return result
-    else:
-        result.update({
-            'form_template': form_definition.form_template_name or settings.DEFAULT_FORM_TEMPLATE
-        })
-        return render_to_response('html/formdefinition/detail.html', result, context_instance=RequestContext(request))
+    result.update({
+        'form_template': form_definition.form_template_name or settings.DEFAULT_FORM_TEMPLATE
+    })
+    return render_to_response('html/formdefinition/detail.html', result,
+        context_instance=RequestContext(request))
